# develop/dbService/src/main/resources/swagger.yml
openapi: 3.0.0
info:
  title: Database API
  version: 1.0.0
paths:
...
  /applications:
    post:
      summary: 'Создание новой заявки'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewApplication'
      responses:
        '201':
          description: 'Заявка создана'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: 'Передан некорректный запрос'
        '500':
          description: 'Внутренняя ошибка сервера'
    get:
      summary: 'Получение всех заявок'
      responses:
        '200':
          description: 'Список всех заявок'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
        '500':
          description: 'Внутренняя ошибка сервера'
  /applications/{id}:
    get:
      summary: 'Получение заявки по ID'
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 'ID заявки'
      responses:
        '200':
          description: 'Заявка найдена'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '404':
          description: 'Заявка не найдена'
        '500':
          description: 'Внутренняя ошибка сервера'
    put:
      summary: 'Обновление заявки'
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 'ID заявки для обновления'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationUpdate'
      responses:
        '200':
          description: 'Заявка обновлена'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: 'Передан некорректный запрос'
        '404':
          description: 'Заявка не найдена'
        '500':
          description: 'Внутренняя ошибка сервера'
...
components:
  schemas:
...
    NewApplication:
      type: object
      required:
        - userId
        - directionId
        - status
      properties:
        userId:
          type: string
          description: 'Идентификатор пользователя'
        directionId:
          type: string
          description: 'Идентификатор направления'
        status:
          type: string
          description: 'Статус заявки'
      example:
        userId: "41a99496-b685-475c-b917-5e956032e034"
        directionId: "3b73ea48-6537-4520-9f14-314bc30f95e6"
        status: "new"
    ApplicationUpdate:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          description: 'Идентификатор заявки'
        userId:
          type: string
          description: 'Идентификатор пользователя'
        directionId:
          type: string
          description: 'Идентификатор направления'
        status:
          type: string
          description: 'Статус заявки'
      example:
        id: 1
        userId: "41a99496-b685-475c-b917-5e956032e034"
        directionId: "3b73ea48-6537-4520-9f14-314bc30f95e6"
        status: "in progress"
    Application:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          description: 'Идентификатор заявки'
        userId:
          type: string
          description: 'Идентификатор пользователя'
        directionId:
          type: string
          description: 'Идентификатор направления'
        status:
          type: string
          description: 'Статус заявки'
      example:
        id: 1
        userId: "41a99496-b685-475c-b917-5e956032e034"
        directionId: "3b73ea48-6537-4520-9f14-314bc30f95e6"
        status: "in progress"
...